remove old components and reorganize components folder
