+++ components/calculator/CTAContainers.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import theme from '../../styles/themes/theme.tsx';
 import StaticCalcProps from '../../utils/calculator.props.ts';
-import { CalculatorButton } from '../helper/CustomButtons.tsx';
+import { CalculatorButton } from '../CustomButtons.tsx';
 
 export default function FinalPageLinksContainer({
   calculatorConfig,
