+++ pages/404.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, Typography } from '@mui/material';
 import React from 'react';
 
-import IndividualPageHead from '../components/helper/IndividualPageHead.tsx';
+import IndividualPageHead from '../components/IndividualPageHead.tsx';
 import ImageContainer from '../components/layout/ImageContainer.tsx';
 import SectionContainer from '../components/layout/SectionContainer.tsx';
 import content from '../content/404.ts';
