+++ pages/get-started/determine-your-eligibility.tsx
@@ -3,8 +3,8 @@ import { Box, Typography } from '@mui/material';
 import MuiMarkdown from 'mui-markdown';
 import React from 'react';
 
-import externalLinks from '../../components/functional/ExternalLinks.tsx';
-import { DarkButton } from '../../components/helper/CustomButtons.tsx';
+import { DarkButton } from '../../components/CustomButtons.tsx';
+import externalLinks from '../../components/ExternalLinks.tsx';
 import GetStartedHero from '../../components/layout/GetStartedHero.tsx';
 import content from '../../content/get-started.ts';
 import { GSContainer } from './index.tsx';
