+++ pages/lfos.tsx
@@ -1,7 +1,7 @@
 import { Typography } from '@mui/material';
 import React from 'react';
 
-import externalLinks from '../components/functional/ExternalLinks.tsx';
+import externalLinks from '../components/ExternalLinks.tsx';
 import AccordionBuilder from '../components/layout/AccordionBuilder.tsx';
 import content from '../content/get-started.ts';
 import { GSContainer } from './get-started/index.tsx';
